moved css and fixed imports
