+++ dashboard/src/pages/Menu.jsx
@@ -16,7 +16,7 @@ import {
 import AddIcon from "@mui/icons-material/Add";
 import RemoveIcon from "@mui/icons-material/Remove";
 import Navbar from "./Navbar";
-import "./Menu.css";
+import "../styles/Menu.css";
 
 const modalStyle = {
     position: "absolute",
