+++ dashboard/src/pages/Sidebar.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link, Navigate } from "react-router-dom";
-import "./Sidebar.css";
+import "../styles/Sidebar.css";
 import logo from "../images/logo.png";
 
 export default function Sidebar() {
