+++ dashboard/src/styles/ManageProducts.css
