+++ dashboard/src/styles/Navbar.css
